requirements.txt
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
pillow==10.2.0
requests==2.31.0
opencv-python==4.9.0.80
scipy==1.12.0
torch==2.2.1
torchinfo==1.8.0
torchvision==0.17.1
tqdm==4.66.2
python-dateutil==2.8.2
s3transfer==0.10.1
botocore==1.34.65
boto3==1.34.65
huggingface_hub==0.21.4
safetensors==0.4.2
roboflow==roboflow-1.1.24
